requirements.txt
descript_audio_codec==1.0.0
fastapi==0.111.0
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.1
Requests==2.32.2
scipy==1.13.1
torch==2.1.2
transformers==4.40.0
uvicorn==0.29.0
